Split template_backend.py into template package
Separate DjustTemplateBackend, DjustTemplate, and serialization
utilities into focused modules. Original file kept as shim.
